\name{stsm.class-package}
\docType{package}
\alias{stsm.class-package}

\title{Class and Methods for Structural Time Series Models}

\description{
This package defines an S4 class for structural time series models 
and provides some basic methods to work with it.
}

\details{
The class defined in this package is used by the package \pkg{stsm} 
as the base framework to implement statistical algorithms 
related to structural time series models.
}

\author{
Javier López-de-Lacalle \email{javlacalle@yahoo.es}
}

\section{References}{
Christophe Genolini. \emph{A (Not So) Short Introduction to S4.
Object Oriented Programming in \R.} V0.5.1. August 20, 2008.
}

\keyword{package, ts}
